CycleGrid
improvments to cycleGrid, subclassing GridBox from SurfaceView instead of CheckBox, and adding backgrounds,  trying to work with creating arrows with SurfaceView.onTouchListener
will probably have to rework how the GridBoxes work, might need to be all part of one surface view, but then unsure of how to interpret the arrows.  perhaps set this aside for now and work on a simpler analysis of cycle by cycle data?
might be able to interpret arrows by general start and stop positions correlating to specific coordinates on the surface view, and then translating that to a cycle