AutoFragment and PostFragment
final updates to the data saving and loading of the AutoFragment and PostFragment parts of the system, and the layouts and data collection for each tab.  Last step towards improved Davis regional system is the teleTab, which will closely resemble the DriveCoachAssistant portion of this app, with a couple of differences.